% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.mdJSON.R
\name{build.mdJSON}
\alias{build.mdJSON}
\title{Build mdJSON Dictionaries}
\usage{
build.mdJSON(x, title)
}
\arguments{
\item{x}{Data frame}

\item{title}{String}
}
\value{
R list
}
\description{
Transforms a tabular data dictionary into an R list that can be subsequently converted to mdJSON and imported to mdEditor as a Dictionary record. The input data frame must be formatted to a \href{https://github.com/hdvincelette/mdJSONdictio/blob/master/inst/templates/build.mdJSON_Dictionary_Template.xlsx?raw=true}{template}.
}
\examples{
# Import tabular data dictionary as data frame
path<-system.file("extdata", "e.g.dictionary.xlsx", package = "mdJSONdictio")
e.g.dictionary<-readxl::read_excel(path)

# Transform data frame to R list
newjson<- build.mdJSON(x = e.g.dictionary, title = "Example Dictionary")

# Convert R list to JSON
e.g.dictionary = rjson::toJSON(newjson)

# Export JSON to disk
write(e.g.dictionary, "e.g.dictionary.json")
}
\keyword{dictionary,}
\keyword{json,}
\keyword{mdEditor,}
\keyword{mdJSON,}
\keyword{metadata}
