% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.mdJSON.R
\name{edit.build.mdJSON}
\alias{edit.build.mdJSON}
\title{Build mdJSON Data Dictionaries}
\usage{
\method{edit}{build.mdJSON}(x, title)
}
\arguments{
\item{x}{A data frame of the tabular data dictionary.}

\item{title}{A string designating the title of the Dictionary record in mdEditor.}
}
\value{
Returns an R list corresponding to the tabular data dictionary.
}
\description{
Transforms a tabular data dictionary (read as a data frame) into an R list that can be subsequently converted to mdJSON and imported to mdEditor as a Dictionary record. The input data frame must be formatted to a \href{https://github.com/hdvincelette/mdJSONdictio/blob/master/inst/templates/build.mdJSON_Dictionary_Template.xlsx?raw=true}{template}.
}
\examples{
# Import tabular data dictionary as data frame
path<-system.file("extdata", "e.g.dictionary.xlsx", package = "mdJSONdictio")
e.g.dictionary<-readxl::read_excel(path)

# Transform data frame to R list
newjson2<- build.mdJSON(x = e.g.dictionary, title = "Example Dictionary")

# Convert R list to JSON
e.g.dictionary = rjson::toJSON(newjson2)

# Export JSON to disk
write(e.g.dictionary, "e.g.dictionary.json")
}
\seealso{
```build.table()````
}
\keyword{dictionary,}
\keyword{json,}
\keyword{mdEditor,}
\keyword{mdJSON,}
\keyword{metadata}
