% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.mdJSON.R
\name{extract.mdJSON}
\alias{extract.mdJSON}
\title{Extract mdJSON Data Dictionaries from a Metadata file}
\usage{
extract.mdJSON(x, record.type = "dictionaries", multiple = TRUE)
}
\arguments{
\item{x}{List object converted from an mdJSON file.}

\item{record.type}{Default="dictionaries". String or vector representing the type(s) of records to extract. Additional options include "records" and "contacts".}
}
\value{
Returns a list object corresponding to the mdJSON data dictionary file.
}
\description{
Plucks one or more mdJSON data dictionaries from an mdJSON file comprised of other types of records.
}
\seealso{
\code{extract.mdJSON()}
}
\keyword{dictionary,}
\keyword{json,}
\keyword{mdEditor,}
\keyword{mdJSON,}
\keyword{metadata}
