% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.mdJSON.R
\name{validate.mdJSON}
\alias{validate.mdJSON}
\title{Validate mdJSON Data Dictionaries}
\usage{
validate.mdJSON(x, y, dictionary_num = 1, entity_num = 1)
}
\arguments{
\item{x}{List object converted from an mdJSON data dictionary file.}

\item{y}{Data frame of a dataset.}

\item{dictionary_num}{Default=1. Integer indicating the dictionary if there is more than one in the mdJSON file (i.e. if multiple dictionaries are exported together in mdEditor).}

\item{entity_num}{Default=1. Integer indicating the entity if there is more than one in the mdJSON file.}
}
\value{
Returns a data frame comprised of warning messages about and the mdJSON data dictionary.
}
\description{
Compares an mdJSON data dictionary to a tabular dataset and summarizes discrepancies in a data frame.
}
\examples{
# Import mdJSON data dictionary as list
path<-system.file("extdata", "e.g.dictionary2.json", package = "mdJSONdictio")
input.dict <- rjson::fromJSON(file = path)

# Import tabular dataset as data frame
path<-system.file("extdata", "e.g.dataset.csv", package = "mdJSONdictio")
input.data<-read.csv(file = path, na.strings = "", stringsAsFactors = FALSE)

# Validate list against data frame
all.warnings<- validate.mdJSON(x = input.dict, y = input.data, dictionary_num = 1, entity_num = 1)

# Export table to disk
write.csv(x = all.warnings, file = "e.g.warnings.csv")
}
\seealso{
\code{validate.mdJSON()}
}
\keyword{dictionary,}
\keyword{json,}
\keyword{mdEditor,}
\keyword{mdJSON,}
\keyword{metadata}
