% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.table.R
\name{validate.table}
\alias{validate.table}
\title{Validate Tabular Data Dictionaries}
\usage{
validate.table(x, y)
}
\arguments{
\item{x}{A data frame of the tabular data dictionary.}

\item{y}{A data frame of the dataset.}
}
\value{

}
\description{
Validate Tabular Data Dictionaries
}
\examples{
#
}
\seealso{
\code{validate.table()}
}
\keyword{dictionary,}
\keyword{json,}
\keyword{mdEditor,}
\keyword{mdJSON,}
\keyword{metadata}
