% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.mdJSON.R
\name{write.mdJSON}
\alias{write.mdJSON}
\title{Write mdJSON Dictionary}
\usage{
write.mdJSON(file, title)
}
\arguments{
\item{file}{A tabular dictionary file.}

\item{title}{A string designating the title of the mdJSON dictionary.}
}
\value{
An mdJSON dictionary file.
}
\description{
Write an mdJSON dictionary from a \href{https://github.com/hdvincelette/mdJSONdictio/blob/master/inst/templates/write.mdJSON_Dictionary_Template.xlsx?raw=true}{tabular dictionary template}. The mdJSON dictionary can be imported to mdEditor to create a Dictionary record.
}
\examples{
f<- system.file("extdata", "e.g.dictionary.xlsx", package = "mdJSONdictio")
write.mdJSON(file = f, title = "Example Dictionary")
}
\keyword{dictionary,}
\keyword{json,}
\keyword{mdEditor,}
\keyword{mdJSON,}
\keyword{metadata}
