% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.mdJSON.R
\name{build.mdJSON}
\alias{build.mdJSON}
\title{Build mdJSON Data Dictionaries}
\usage{
build.mdJSON(x, title)
}
\arguments{
\item{x}{A data frame of the tabular data dictionary.}

\item{title}{A string designating the title of the Dictionary Record in mdEditor.}
}
\value{
Returns a list object corresponding to the tabular data dictionary.
}
\description{
Translates a data frame of a tabular data dictionary into an list object that can be subsequently converted to mdJSON and imported to mdEditor as a Dictionary Record. The input tabular data dictionary must be formatted to a \href{https://github.com/hdvincelette/mdJSONdictio/blob/master/inst/templates/mdJSONdictio_Dictionary_Template_v1.xlsx?raw=true}{template}.
}
\examples{
# Import tabular data dictionary as data frame
path<-system.file("extdata", "e.g.dictionary.xlsx", package = "mdJSONdictio")
input.table<-readxl::read_excel(path = path)

# Translate data frame to list
new.list<- mdJSONdictio::build.mdJSON(x = input.table, title = "Example Dictionary")

# Convert list to JSON
new.json = rjson::toJSON(x = new.list)

# Export JSON to disk
write(x = new.json, file = "e.g.dictionary.json")
}
\seealso{
\code{build.table()}
}
\keyword{dictionary,}
\keyword{json,}
\keyword{mdEditor,}
\keyword{mdJSON,}
\keyword{metadata}
