% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify.mdJSON.R
\name{modify.mdJSON}
\alias{modify.mdJSON}
\title{Modify mdJSON Data Dictionaries}
\usage{
modify.mdJSON(
  x,
  how = c("add_attribute", "add_domain", "add_domainItem", "update_attribute"),
  codeName
)
}
\arguments{
\item{x}{List object converted from an mdJSON file.}

\item{how}{character string matching one of the modification options: see ‘Details’.}

\item{codeName}{a string representing the attribute to add or modify.}
}
\value{
Returns a modified list object corresponding to the mdJSON data dictionary file.
}
\description{
Amends an mdJSON data dictionary, including adding/updating attributes and domains.
}
\examples{
# Import mdJSON data dictionary as list
path<-system.file("extdata", "e.g.dictionary2.json", package = "mdJSONdictio")
input.dxnry <- rjson::fromJSON(file = path)

# Add an attribute
modified.dxnry<- modify.mdJSON(x = input.dxnry, how = "add_attribute", codeName = "WingArea")

#' # Add a domain
modified.dxnry<- modify.mdJSON(x = input.dxnry, how = "add_domain")

# Convert list to JSON
new.json = rjson::toJSON(x = modified.dxnry)

# Export JSON to disk
write(x = new.json, file = "e.g.dictionary.json")
}
\seealso{
\code{modify.mdJSON()}
}
\keyword{dictionary,}
\keyword{json,}
\keyword{mdEditor,}
\keyword{mdJSON,}
\keyword{metadata}
