% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.table.R
\name{validate.table}
\alias{validate.table}
\title{Validate Tabular Data Dictionaries}
\usage{
validate.table(x, y)
}
\arguments{
\item{x}{A data frame of the tabular data dictionary.}

\item{y}{A data frame of the dataset.}
}
\value{

}
\description{
Validate Tabular Data Dictionaries
}
\examples{
# Import tabular data dictionary as data frame
path<-system.file("extdata", "e.g.dictionary.xlsx", package = "mdJSONdictio")
input.dict<-readxl::read_excel(path = path)

# Import dataset as data frame
path<-system.file("extdata", "e.g.dataset.csv", package = "mdJSONdictio")
input.data<-read.csv(path, na.strings = "", stringsAsFactors = FALSE)

# Validate dataset against dictionary
all.warnings<- validate.table(input.dict, input.data)

# Export table to disk
write.csv(x = all.warnings, file = "e.g.warnings.json")
}
\seealso{
\code{validate.table()}
}
\keyword{dictionary,}
\keyword{json,}
\keyword{mdEditor,}
\keyword{mdJSON,}
\keyword{metadata}
