% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.table.R
\name{validate.table}
\alias{validate.table}
\title{Validate Tabular Data Dictionaries}
\usage{
validate.table(x, y)
}
\arguments{
\item{x}{Data frame of a tabular data dictionary.}

\item{y}{Data frame of a dataset.}
}
\value{
Returns a data frame comprised of warning messages about the tabular data dictionary.
}
\description{
Compares a tabular data dictionary against a dataset and summarizes discrepancies in a data frame. The input data dictionary must be formatted to a \href{https://github.com/hdvincelette/mdJSONdictio/blob/master/inst/templates/mdJSONdictio_Dictionary_Template_v1.xlsx?raw=true}{template}.
}
\examples{
# Import tabular data dictionary as data frame
path<-system.file("extdata", "e.g.dictionary.xlsx", package = "mdJSONdictio")
input.dict<-readxl::read_excel(path = path)

# Import dataset as data frame
path<-system.file("extdata", "e.g.dataset.csv", package = "mdJSONdictio")
input.data<-read.csv(path, na.strings = "", stringsAsFactors = FALSE)

# Validate dataset against dictionary
all.warnings<- validate.table(input.dict, input.data)

# Export table to disk
write.csv(x = all.warnings, file = "e.g.warnings2.json")
}
\seealso{
\code{validate.table()}
}
\keyword{dictionary,}
\keyword{json,}
\keyword{mdEditor,}
\keyword{mdJSON,}
\keyword{metadata}
